+++ src/main/java/site/hirecruit/hr/domain/auth/service/impl/OAuth2ProcessorFacadeImpl.kt
@@ -1,8 +1,11 @@
-package site.hirecruit.hr.domain.auth.service
+package site.hirecruit.hr.domain.auth.service.impl
 
 import org.springframework.stereotype.Service
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
 import site.hirecruit.hr.domain.auth.model.User
+import site.hirecruit.hr.domain.auth.service.OAuthProcessorFacade
+import site.hirecruit.hr.domain.auth.service.UserAuthService
+import site.hirecruit.hr.domain.auth.service.UserRegistrationService
 import site.hirecruit.hr.domain.worker.repository.WorkerRepository
 
 /**
