+++ src/main/java/site/hirecruit/hr/domain/auth/service/impl/UserAuthServiceImpl.kt
@@ -1,8 +1,9 @@
-package site.hirecruit.hr.domain.auth.service
+package site.hirecruit.hr.domain.auth.service.impl
 
 import org.springframework.stereotype.Service
 import site.hirecruit.hr.domain.auth.dto.OAuthAttributes
 import site.hirecruit.hr.domain.auth.model.User
+import site.hirecruit.hr.domain.auth.service.UserAuthService
 
 /**
  * 세션기반 인증을 진행하는 서비스
